+++ components/tas58xx/tas58xx_helpers.h
@@ -14,6 +14,6 @@ struct BiquadCoefficients {
 
 int32_t gain_to_f9_23_(int8_t gain);
 
-BiquadCoefficients equalizer_qfactor_calc(uint32_t sample_rate, float frequency, int16_t gain, float qFactor);
+BiquadCoefficients equalizer_qfactor_calc(uint32_t sample_rate, uint16_t frequency, int16_t gain, float qFactor);
 
 }  // namespace esphome::tas58xx_helpers
